+++ src/components/Header/components/MobileNav.tsx
@@ -11,7 +11,7 @@ import Link from "../../Link/Link";
 import List from "../../List/List";
 import Logo from "../../Logo/Logo";
 import SimpleGrid from "../../Grid/SimpleGrid";
-import { upperNavLinks, lowerNavLinks } from "../headerUtils";
+import { upperNavLinks, lowerNavLinks } from "../utils/headerUtils";
 
 /**
  * The Header navigation for the mobile view.