+++ src/theme/components/header/headerSearchButton.ts
@@ -1,18 +1,21 @@
 import { headerFocus } from "./header";
 
 const HeaderSearchButton = {
-  baseStyle: ({ isMobile, isOpen }) => ({
+  baseStyle: ({ isOpen }) => ({
     alignItems: "center",
     borderRadius: "0",
     backgroundColor: isOpen ? "#1B7FA7" : "transparent",
     color: isOpen ? "ui.white" : null,
     fontSize: "inherit",
     fontWeight: 500,
-    minHeight: isMobile ? "50px" : null,
-    minWidth: isMobile ? "50px" : "80px",
+    minHeight: { base: "50px", md: "40px" },
+    minWidth: { base: "50px", md: "80px" },
     svg: {
-      marginLeft: isMobile ? "0" : null,
-      color: isOpen ? "ui.white" : isMobile ? "ui.black" : "ui.link.primary",
+      marginLeft: { base: "0", md: null },
+      color: {
+        base: isOpen ? "ui.white" : "ui.black",
+        md: isOpen ? "ui.white" : "ui.link.primary",
+      },
     },
     textDecoration: isOpen ? "none" : "underline",
     _hover: {