+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates to Storybook version 6.5.
 - Explicitly sets the default color mode value to `"light"`.
 - Updates how the `styles.scss` and `resources.scss` files are organized and compiled so that they can be imported in any tech stack.
+- Updates the docs for the `Table` commponent to remove the example that does not include column headers.
 
 ## 1.0.3 (June 9, 2022)