+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Updates how the `styles.scss` and `resources.scss` files are organized and compiled so that they can be imported in any tech stack.
 - Updates the docs for the `Table` commponent to remove the example that does not include column headers.
 - Updates the `Button`, `Select` and `TextInput` components to use NYPL standard minimum height in mobile viewport.
+- Updates the `SearchBar` component to change how the width of the internal `Select` component is handled.
 
 ## 1.0.3 (June 9, 2022)